auth/utils/jaeger.py
from core.config import TestSettings
from opentelemetry import trace
from opentelemetry.exporter.jaeger.thrift import JaegerExporter
from opentelemetry.sdk.resources import SERVICE_NAME, Resource
from opentelemetry.sdk.trace import TracerProvider
from opentelemetry.sdk.trace.export import BatchSpanProcessor

settings = TestSettings()

resource = Resource(attributes={
    SERVICE_NAME: "users_service"
})


provider = TracerProvider(resource=resource)
processor = BatchSpanProcessor(JaegerExporter(
    agent_host_name=settings.JAEGER_HOST,
    agent_port=int(settings.JAEGER_PORT),
))
provider.add_span_processor(processor)
trace.set_tracer_provider(provider)


tracer = trace.get_tracer(__name__)
