auth/requirements.txt
Flask==2.2.2
gevent==22.10.1
sqlalchemy==1.4.42
psycopg2-binary==2.9.5
requests==2.28.1
redis==4.1.0
aiohttp==3.7.2
flask_sqlalchemy==3.0.2
flask-marshmallow==0.14.0
flask-apispec==0.11.4
flask-swagger-ui==4.11.1
apispec-webframeworks==0.5.2
flasgger==0.9.5
email_validator==1.3.0
Flask-WTF==1.0.1
flask-restx==0.5.1
Flask-Migrate==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-praetorian==1.4.0
Flask-RESTful==0.3.9
passlib==1.7.4
Flask-JWT-Extended==4.4.4
email-validator==1.3.0
password-validator==1.0
flask-migrate==3.1.0
flask_script==2.0.6
python-dotenv==0.21.0
pydantic==1.10.2
Authlib==1.0.0
transliterate==1.10.2

opentelemetry-distro==0.35b0
opentelemetry-exporter-otlp==1.14.0
opentelemetry-api==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-instrumentation-flask==0.35b0
opentelemetry-exporter-jaeger==1.14.0

request-id-flask==0.0.2
flask-request-id-header==0.1.1
opentelemetry-instrumentation-requests==0.35b0
Flask-OpenTracing==1.1.0

Flask-Limiter==2.7.0

ua-parser==0.16.1
user-agents==2.2.0
captcha==0.4
Random-Word==1.0.11
circuitbreaker==1.4.0
